created BEN
we have made thhe one and only ben who was made to rule them all and can turn the wearer invisible. Ben will also tempt the user to return to Sauron so they can become god but if the user is a hobbit they won't give enough fucks to try.

Wait that was about the one ring ben is just a person who took part in making this game